+++ src/main/java/org/openwebnet4j/communication/BUSConnector.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020 Contributors to the openwebnet4j project
+ * Copyright (c) 2021 Contributors to the openwebnet4j project
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information.
@@ -307,7 +307,7 @@ public class BUSConnector extends OpenConnector {
             // STEP-2: NO_AUTH - Free beer and party, the connection is unauthenticated!
             frCh.handshakeCompleted = true;
             hsLogger.debug("(HS) ... STEP-2: NO_AUTH: second ACK received, GW has no pwd ==HANDSHAKE COMPLETED==");
-        } else if (fr.matches("\\*#\\d{7,12}##")) {
+        } else if (fr.matches("\\*#\\d+##")) {
             // STEP-2: OPEN_AUTH passwd nonce received
             doOPENHandshake(fr, frCh);
             frCh.handshakeCompleted = true;