+++ src/main/java/org/openwebnet4j/communication/FrameChannel.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020 Contributors to the openwebnet4j project
+ * Copyright (c) 2021 Contributors to the openwebnet4j project
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information.
@@ -135,7 +135,7 @@ public class FrameChannel {
     /**
      * Reads from InputStream until delimiter ('##'), putting data into buffer
      *
-     * @returns number of bytes read, or -1 in case of end of stream
+     * @return number of bytes read, or -1 in case of end of stream
      * @throws IOException in case of problems with the InputStream
      */
     private int readUntilDelimiter(InputStream is, byte[] buffer) throws IOException {