+++ src/main/java/org/openwebnet4j/communication/USBConnector.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020 Contributors to the openwebnet4j project
+ * Copyright (c) 2021 Contributors to the openwebnet4j project
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information.
@@ -420,7 +420,7 @@ public class USBConnector extends OpenConnector implements SerialPortEventListen
      * Use this instead of String.compareTo() for a non-lexicographical
      * comparison that works for version strings. e.g. "1.10".compareTo("1.6").
      *
-     * @note It does not work if "1.10" is supposed to be equal to "1.10.0".
+     * @apiNote It does not work if "1.10" is supposed to be equal to "1.10.0".
      *
      * @param str1 a string of ordinal numbers separated by decimal points.
      * @param str2 a string of ordinal numbers separated by decimal points.