+++ src/main/java/org/openwebnet4j/message/GatewayMgmt.java
@@ -1,5 +1,5 @@
 /**
- * Copyright (c) 2020 Contributors to the openwebnet4j project
+ * Copyright (c) 2021 Contributors to the openwebnet4j project
  *
  * See the NOTICE file(s) distributed with this work for additional
  * information.
@@ -132,7 +132,7 @@ public class GatewayMgmt extends BaseOpenMessage {
     }
 
     /**
-     * OpenWebNet message request for supervisor mode <b>*13*66*##</b>.
+     * OpenWebNet message request for supervisor mode <code>*13*66*##</code>.
      *
      * @return GatewayMgmt message
      */
@@ -141,7 +141,7 @@ public class GatewayMgmt extends BaseOpenMessage {
     }
 
     /**
-     * OpenWebNet message request keep connect <b>*13*60*##</b>.
+     * OpenWebNet message request keep connect <code>*13*60*##</code>.
      *
      * @return GatewayMgmt message
      */
@@ -150,18 +150,20 @@ public class GatewayMgmt extends BaseOpenMessage {
     }
 
     /**
-     * OpenWebNet message request for gateway MAC address <b>*#13**12##</b>.
+     * OpenWebNet message request for gateway MAC address <code>*#13**12##</code>.
      *
      * @return GatewayMgmt message
      */
     public static GatewayMgmt requestMACAddress() {
-        return new GatewayMgmt(format(FORMAT_DIMENSION, WHO, "", DIM.MAC_ADDRESS.value()));
+        return new GatewayMgmt(format(FORMAT_DIMENSION_REQUEST, WHO, "", DIM.MAC_ADDRESS.value()));
     }
 
     /**
      * Parse MAC address in the OWN message and return values in byte[]
      *
+     * @param msg the message to parse
      * @return byte[] MAC address values
+     * @throws FrameException in case of error in frame
      */
     public static byte[] parseMACAddress(GatewayMgmt msg) throws FrameException {
         // MAC address is returned in VAL1-VAL6 or VAL1-VAL8 decimal dimensions of the MAC address response frame
@@ -174,21 +176,21 @@ public class GatewayMgmt extends BaseOpenMessage {
     }
 
     /**
-     * OpenWebNet message request for gateway model <b>*#13**15##</b>.
+     * OpenWebNet message request for gateway model <code>*#13**15##</code>.
      *
      * @return GatewayMgmt message
      */
     public static GatewayMgmt requestModel() {
-        return new GatewayMgmt(format(FORMAT_DIMENSION, WHO, "", DIM.MODEL.value()));
+        return new GatewayMgmt(format(FORMAT_DIMENSION_REQUEST, WHO, "", DIM.MODEL.value()));
     }
 
     /**
-     * OpenWebNet message request for gateway firmware version <b>*#13**16##</b>.
+     * OpenWebNet message request for gateway firmware version <code>*#13**16##</code>.
      *
      * @return GatewayMgmt message
      */
     public static GatewayMgmt requestFirmwareVersion() {
-        return new GatewayMgmt(format(FORMAT_DIMENSION, WHO, "", DIM.FIRMWARE_VERSION.value()));
+        return new GatewayMgmt(format(FORMAT_DIMENSION_REQUEST, WHO, "", DIM.FIRMWARE_VERSION.value()));
     }
 
     public static String parseFirmwareVersion(GatewayMgmt msg) throws FrameException {
@@ -198,7 +200,7 @@ public class GatewayMgmt extends BaseOpenMessage {
     }
 
     /**
-     * OpenWebNet message request to scan network <b>*13*65*##</b>.
+     * OpenWebNet message request to scan network <code>*13*65*##</code>.
      *
      * @return GatewayMgmt message
      */
@@ -207,30 +209,21 @@ public class GatewayMgmt extends BaseOpenMessage {
     }
 
     /**
-     * OpenWebNet message request for product information <b>*#13**66*index##</b>.
+     * OpenWebNet message request for product information <code>*#13**66*index##</code>.
+     *
+     * <p>
+     * <b>NOTE</b>
+     * Due to a bug in the USB gateway, request product info message must use <code>*</code> to separate index instead
+     * of <code>#</code>: <code>*#13**66*index##</code> instead of <code>*#13**66#index##</code> as documented in
+     * OpenWebNet specs.
      *
      * @param index The index of the product inside the gateway products database as returned from network scan. Index
      *            starts at 0.
      * @return GatewayMgmt message
-     *
-     * @note Due to a bug in the USB gateway, request product info message must use '*' to separate index instead of
-     *       '#'.
-     *
-     *       <pre>
-     * *#13**66<b>*</b>index##
-     *       </pre>
-     *
-     *       instead of
-     *
-     *       <pre>
-     * *#13**66<b>#</b>index##
-     *       </pre>
-     *
-     *       as documented in OpenWebNet specs
      */
     public static GatewayMgmt requestProductInfo(int index) {
         // we must use here addValues instead of addDimensions to be compatible with gateway bug
-        String req = format(FORMAT_DIMENSION, WHO, "", DIM.PRODUCT_INFO.value());
+        String req = format(FORMAT_DIMENSION_REQUEST, WHO, "", DIM.PRODUCT_INFO.value());
         req = addValues(req, index + "");
         return new GatewayMgmt(req);
     }